+++ inc/dbconnection.class.php
@@ -573,9 +573,9 @@ class DBConnection extends CommonDBTM {
 
       // No need to translate, this part always display in english (for copy/paste to forum)
 
-      echo "<tr class='tab_bg_2'><th>".self::getTypeName(Session::getPluralNumber())."</th></tr>";
+      echo "<tr class='tab_bg_2'><th class='section-header'>".self::getTypeName(Session::getPluralNumber())."</th></tr>";
 
-      echo "<tr class='tab_bg_1'><td><pre>\n&nbsp;\n";
+      echo "<tr class='tab_bg_1'><td><pre class='section-content'>\n&nbsp;\n";
       if (self::isDBSlaveActive()) {
          echo "Active\n";
          self::showAllReplicateDelay();