+++ js/common.js
@@ -1330,3 +1330,45 @@ $(
 jQuery.expr.filters.icontains = function(elem, i, m) {
    return (elem.innerText || elem.textContent || "").toLowerCase().indexOf(m[3].toLowerCase()) > -1;
 };
+
+function tableToDetails(table) {
+   let in_details = false;
+   const section_els = $(table).find('.section-header, .section-content');
+   let details = '';
+
+   section_els.each((i, e) => {
+      if (e.classList.contains('section-header')) {
+         if (in_details) {
+            details += '</details>';
+         }
+         details += `<details><summary>${e.innerText}</summary><pre>`;
+         in_details = true;
+      } else {
+         if (in_details) {
+            details += e.innerText;
+         }
+      }
+   });
+
+   if (in_details) {
+      details += '</pre></details>';
+   }
+   return details;
+}
+
+function flashIconButton(button, button_classes, icon_classes, duration) {
+   const btn = $(button);
+   const ico = btn.find('i').eq(0);
+   const original_btn_classes = btn.attr('class');
+   const original_ico_classes = ico.attr('class');
+   btn.removeClass();
+   ico.removeClass();
+   btn.addClass(button_classes);
+   ico.addClass(icon_classes);
+   window.setTimeout(() => {
+      btn.removeClass();
+      ico.removeClass();
+      btn.addClass(original_btn_classes);
+      ico.addClass(original_ico_classes);
+   }, duration);
+}